Setup input for our RightClick key binding in our TDMPlayerController.
Added setup for handing when the player presses and releases the "RightClick" key binding which is Right Mouse Button by default. Input is only handled on the client so on Tick we will check if the right mouse button is down and do a RPC for the server to handle it.