Added two Trace Channels in the projects settings: Floor, and Targetable.
We need two trace channels so we can use one for skill shot abilities where we don't want other characters to effect the location. We need this because the camera isn't directly above the player and at the angle we will use, the character can cause weird targeting problems so we use a dedicated trace channel for that where the Floor is the only actor that will trigger it.